requirements.txt
appnope==0.1.0
backports-abc==0.5
cffi==1.9.1
click==6.7
conda==4.3.13
cookies==2.2.1
cryptography==1.7.1
decorator==4.0.11
facebook-sdk==2.0.0
Flask==0.12
geopy==1.11.0
googlefinance==0.7
httplib2==0.10.3
idna==2.2
ipykernel==4.5.2
ipython==5.1.0
ipython-genutils==0.1.0
ipywidgets==5.2.2
itsdangerous==0.24
Jinja2==2.9.4
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.4.0
jupyter-console==5.0.0
jupyter-core==4.2.1
MarkupSafe==0.23
mistune==0.7.3
nbconvert==4.2.0
nbformat==4.2.0
notebook==4.3.1
numpy==1.9.2
oauth2==1.9.0.post1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pyasn1==0.1.9
pycosat==0.6.1
pycparser==2.17
pycrypto==2.6.1
Pygments==2.1.3
pyOpenSSL==16.2.0
python-dotenv==0.6.3
python-forecastio==1.3.5
PyYAML==3.11
pyzmq==16.0.2
qtconsole==4.2.1
requests==2.12.4
responses==0.5.1
simplegeneric==0.8.1
six==1.10.0
terminado==0.6
tornado==4.4.2
traitlets==4.3.1
wcwidth==0.1.7
Werkzeug==0.11.15
widgetsnbextension==1.2.6
yelp==1.0.2
